Small improvements

+++ number_obfuscate/Node/ObfuscateConfig.h
@@ -8,8 +8,8 @@ struct ObfuscateConfig {
     std::string unknownVariableName = "j";
     int64_t knownVariableValue = 5;
     uint8_t deep = 5;
-    const int64_t min = INT8_MIN;
-    const int64_t max = INT8_MAX;
+    const int64_t min = INT16_MIN;
+    const int64_t max = INT16_MAX;
 
     ObfuscateConfig() {};
     explicit ObfuscateConfig(const uint8_t deepVal): deep(deepVal) {}
